requirements.txt
arxiv
requests
pypdf
fpdf
google-adk
Flask